% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{italian_legislator_17}
\alias{italian_legislator_17}
\title{Italian legislator 17}
\format{
An object of class \code{list} of length 4.
}
\usage{
italian_legislator_17
}
\description{
Italian legislator 17
}
\keyword{datasets}
