% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{italian_legislator_18}
\alias{italian_legislator_18}
\title{Italian legislator 18}
\format{
An object of class \code{list} of length 4.
}
\usage{
italian_legislator_18
}
\description{
Italian legislator 18
}
\keyword{datasets}
