% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{swiss_legislator_49}
\alias{swiss_legislator_49}
\title{Swiss legislator 49}
\format{
An object of class \code{list} of length 4.
}
\usage{
swiss_legislator_49
}
\description{
Full data for the swiss legislative period 17.
}
\keyword{datasets}
