% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{swiss_legislator_50}
\alias{swiss_legislator_50}
\title{Swiss legislator 50}
\format{
An object of class \code{list} of length 4.
}
\usage{
swiss_legislator_50
}
\description{
Swiss legislator 50
}
\keyword{datasets}
