% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{italian_legislator_17}
\alias{italian_legislator_17}
\title{Italian legislator 17}
\format{
A list of 3 tibbles
\describe{
\item{\code{members_of_parliment}}{\describe{
\item{\code{i_id}}{an identifier}
\item{\code{uri}}{the link to the online resource about the person}
\item{\code{first_name}}{First Name}
\item{\code{last_name}}{Surname}
\item{\code{info}}{Diplomas and work information}
\item{\code{birth_date}}{Date of birth}
\item{\code{death_date}}{Date of death}
\item{\code{gender}}{gender}
\item{\code{party}}{\describe{
\item{\code{party_name}}{Name of the parlamentary group}
\item{\code{start_date}}{Affiliation start date}
\item{\code{end_date}}{Affiliation end date}
}}
\item{\code{election_date}}{date of the election}
\item{\code{electoral_district}}{The region represented by the member}
\item{\code{election_list}}{Name of the electoral list}
\item{\code{start_mandate}}{Start date of the mandate}
\item{\code{end_mandate}}{End date of the mandate}
\item{\code{end_motive}}{Motive of the mandate termination}
\item{\code{region}}{The region represented by the member}
\item{\code{has_scores}}{Indicates if the councillor has score information}
}}
\item{\code{voting_items}}{\describe{
\item{\code{v_id}}{an identifier}
\item{\code{date}}{The Vote date}
\item{\code{type}}{The vote type}
\item{\code{act_id}}{The associated act id}
\item{\code{act_type}}{The type of act}
\item{\code{act_demanded_by}}{The demanding autority of the act}
\item{\code{description}}{The description of the item}
\item{\code{final_vote}}{Is a final vote}
\item{\code{trust_vote}}{Is a trust vote}
\item{\code{secret_vote}}{Is a secret vote}
}}
\item{\code{polls}}{The poll results for each \code{member_of_parliment} by each \code{voting_item} in integer form}
\item{\code{poll_codes}}{ The codes of the poll data}
\item{\code{partys}}{\describe{
\item{\code{party_name_long}}{The long party name}
\item{\code{party_name}}{The abbreviated version of the name}
\item{\code{old_names}}{Names the same party has taken over time}
}}
}
}
\usage{
italian_legislator_17
}
\description{
Italian legislator 17
}
\keyword{datasets}
