% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{italian_legislator_17}
\alias{italian_legislator_17}
\title{italian_legislator_17}
\format{
A list containing tbls and a matrix
\describe{
  \item{members_of_parliment}{tbl}
  \item{voting_items}{tbl}
  \item{polls}{poll matrix}
  \item{poll_codes}{numeric codes used in the polls matrix and their meaning}
  \item{partys}{changes of party names}
}
}
\source{
\url{https://dati.camera.it/}
}
\usage{
italian_legislator_17
}
\description{
italian_legislator_17
}
\keyword{datasets}
