% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{swiss_legislator_49}
\alias{swiss_legislator_49}
\title{swiss_legislator_49}
\format{
A list containing tbls and a matrix
\describe{
  \item{members_of_parliment}{tbl}
  \item{voting_items}{tbl}
  \item{polls}{poll matrix}
  \item{poll_codes}{numeric codes used in the polls matrix and their meaning}
}
}
\source{
\url{http://ws-old.parlament.ch/}

\url{https://www.parlament.ch/de/ratsbetrieb/abstimmungen/abstimmung-nr-xls}
}
\usage{
swiss_legislator_49
}
\description{
swiss_legislator_49
}
\keyword{datasets}
