% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{swiss_legislator_49}
\alias{swiss_legislator_49}
\title{Swiss legislator 49}
\format{
A list of 3 tibbles
\describe{
\item{\code{members_of_parliment}}{\describe{
\item{\code{i_id}}{an identifier}
\item{\code{uri}}{the link to the online resource about the person}
\item{\code{first_name}}{First Name}
\item{\code{last_name}}{Surname}
\item{\code{birth_date}}{Date of birth}
\item{\code{death_date}}{Date of death}
\item{\code{gender}}{gender}
\item{\code{home_place}}{home_place}
\item{\code{parties}}{\describe{
\item{\code{party_name}}{Name of the parlamentary group}
\item{\code{start_date}}{Affiliation start date}
\item{\code{end_date}}{Affiliation end date}
}}
\item{\code{factions}}{\describe{
\item{\code{faction}}{Name of the parlamentary group}
\item{\code{start_date}}{Affiliation start date}
\item{\code{end_date}}{Affiliation end date}
}}
\item{\code{start_mandate}}{Start date of the mandate}
\item{\code{end_mandate}}{End date of the mandate}
\item{\code{canton}}{The region represented by the member}
\item{\code{has_scores}}{Indicates if the councillor has score information}
}}
\item{\code{voting_items}}{\describe{
\item{\code{v_id}}{an identifier}
\item{\code{session_code}}{the session}
\item{\code{vote_date}}{The Vote date}
\item{\code{committee}}{The committee}
\item{\code{concernes_department}}{The concernes_department}
\item{\code{affair_id}}{The affair_id}
\item{\code{affair_title}}{The affair_title}
\item{\code{meaning_yes}}{The meaning_yes}
\item{\code{meaning_no}}{meaning_no}
\item{\code{vote_type}}{vote_type}
\item{\code{vote_text}}{vote_text}
}}
\item{\code{polls}}{The poll results for each \code{member_of_parliment} by each \code{voting_item} in integer form}
\item{\code{poll_codes}}{ The codes of the poll data}
\item{\code{partys}}{\describe{
\item{\code{party_name_long}}{The long party name}
\item{\code{party_name}}{The abbreviated version of the name}
\item{\code{old_names}}{Names the same party has taken over time}
}}
}
}
\usage{
swiss_legislator_49
}
\description{
Swiss legislator 49
}
\keyword{datasets}
